Move citation formatter to its own file and use language file (#868)
